+++ app/models/artist.rb
@@ -3,6 +3,5 @@ class Artist
 
   property :name, type: String
   property :popularity, type: Integer
-  
   has_many :in, :users, origin: :artists
 end
