+++ app/models/user.rb
@@ -3,6 +3,5 @@ class User
 
   property :username, type: String
   property :email, type: String
-
   has_many :out, :artists, type: :LISTENS_TO
 end
