+++ app/views/artists/match.html.erb
@@ -13,5 +13,5 @@
   <div id="intro_text">
     Click below to
   </div>
-  <%= link_to 'generate a playlist', '/users/playlist', class:"link", id:"generate"%>
+  <%= link_to 'generate a playlist', '/artists/playlist', class:"link", id:"generate"%>
 </div>
