+++ app/views/artists/playlist.html.erb
