+++ spec/controllers/artists_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe ArtistsController, type: :controller do
+
+end
